Updated the code to handle problems centered around numpy.histogram in different versions of NumPy. The issue is that the default behavior of numpy.histogram changed in NumPy 1.2.0. To get consistent results between NumPy 1.2.0 or later, and versions prior to NumPy 1.2.0, new=True must be passed to histgram() in Numpy < 1.2.0.  The following describes the behavior of histogram based on different versions of NumPy, where the 'correct result' is defined as 'the result from NumPy 1.2.0 or later'.
NumPy version less than 1.2.0:
 pass new=True: correct result
 don't pass new=True: incorrect result, warning about different behavior

NumPy version 1.2.0 through 1.3.*:
 pass new=True: correct result, warning about new=True not being necessary
 don't pass new=True: correct result

NumPy version 1.4.0 or later:
 pass new=True: TypeError, no such parameter as new (deprecated)
 don't pass new=True: correct result


We ideally want to have no warning message and the correct result, regardless of numpy version. To achieve that, I've added an explicit check for the NumPy version number, and when version is less than 1.2.0, histrogram is overwritten to include new=True as the default. This allows for no changes in the make_histrogram function. If numpy.histogram is used in other parts of Qiime, we may want to overwrite histogram in qiime.util to take care of this behavior, and then import histogram from there.