Updated all parallel scripts to make use of the new polling system, which monitors jobs running in parallel on the cluster via the existance of their output files. Currently each parallel script is set to use the poller by default, which can be overwritten by passing -W.
When a parallel script runs with the poller, one additional job is launched which knows how to monitor the other jobs for completion. When the jobs finish, the poller runs a function (which can be passed to the poller via the command line) which joins like output files into a single file so that the results from running the serial or parallel versions of the scripts are the same. Additionally, the poller will clean up all temp files (but currently leaves the queueing system's stdout/stderr in the user's home directory). Cleaning up of temp files can be suppressed by passing -R. 

I additionally modified parallel scripts to build temp blast databases (where applicable) before tasks are farmed out, now that they can be cleaned-up via the poller. This is more disk-space-friendly.

I'll be adding documentation to the project docs on the polling system shortly -- once we figure out where that is -- and will be attempting to define a class TaskFarmer which individual task farmers (currently the individual parallel scripts) can inherit from as there is currently quite bit of repeated code.