Added a chimera checking script. Currently there is one ChimeraChecker defined here: BlastFragmentsChimeraChecker, which splits an input sequence into n (default: 3, based on results presented in Lahr and Katz (2009), which suggests that greater than 2 is necessary) roughly equal-sized fragments, and assigns the taxonomy of each fragment using the BlastTaxonAssigner. If different assignments are returned for different fragments, the input sequence is considered to be chimeric.
The BlastFragmentsChimeraChecker is an admittedly simple way to do this. The goal for this script is to implement/wrap several chimera checkers to support multiple ways of doing chimera checking, and to define a MetaChimeraChecker which can take several ChimeraCheckers and string them together and return intersections/unions of the individual ChimeraCheckers, and allow for bake-off-style comparisons of different chimera checking methodologies.