Added a blast-based OTU picker which will define OTUs as sets of sequences which achieve the same best hit against a reference database. This will be useful, for example, if we want to define OTUs against an existing sequence collection and tree and is one way that we can begin comparing OTUs between analyses. For this reason, OTU ids are assigned based on the seq id of the best blast hit rather than as ascending integers as in the other OTU pickers.
THIS CODE IS AWAITING SOME ADDITIONAL UNIT TESTS, AND SHOULDN'T BE USED BEFORE THOSE ARE ADDED. A message to that effect is printed to stdout upon calling the function, and the otu picker is not yet tied into the command line interface. Both of those indicators will be removed when this is ready to be used. I wanted to get this code committed as it is almost ready, and didn't want to deal with a merge if someone else is working on this file before I get a chance to get back to it.