Added new get_options_lookup() function in qiime/util.py which provides a dict of commonly used options. This should help in the effort to standardize the command line interfaces, and is added in preparation for the documentation sprint when we'll incorporate the use of some of these options.
Also added jobs_to_start and seconds_to_sleep to the qiime_config, as these can be convenient for the users to have control over - particularly jobs_to_start will be useful to modify based on the system (e.g., 4 on a 4 core system; 50 on a big cluster; ...)

Added a function, merge_otu_tables, and corresponding tests to handle merging of OTU tables with compatible OTU ids but different sample ids. This may become more useful as we begin switching to tree-based OTU picker, and currently can be use with different runs of the BLAST OTU picker. 

Modified scripts/make_qiime_py_file.py to include a call to get_options_lookup() and an example of how it can be used.