Updated the blast taxonomy assigner to include the id of the best BLAST hit. This is in response to Sourceforge feature request #2988407.
As this adds new data to the output of assign taxonomy, slight modification had to be made in other places. The assign taxonomy output format now supports addition columns after the confidence score which can be used to include method-specific data. These additional fields are optional, so the format is fully backward compatible.