Added note about translated search/functional assignment code.
Note that the previous commit accidentally contained a lot more code than it was supposed to. That check-in was supposed to only be of qiime/parallel/util.py. The other code that went in is a refactoring of the 'functional assignment' code, which is now more accruately described as database mapping software. Testers of the previous functional_assignment.py and parallel_functional_assignment_usearch.py scripts will notice that those are now called assign_reads_to_database.py and parallel_assign_reads_to_database_usearch.py. Some additional changes: (1) serial version now supports BLAT, and parallel will soon too; (2) serial version no longer supports blastx - this would have taken a while to refactor for the new framework and it was way too slow to be useable, so as of now I don't think it's worth the time; (3) output now includes BIOM tables, which should make this code easier to plug into downstream applications.