modified output format
This is an ugly hack, and needs to be pushed upstream to the taxon assigners. The output taxonomy maps that are returned by the  taxon assigners contain the full sequence headers as the first field (so including "comment" text in the fasta headers), but for consistency with the input taxonomy maps, should only contain the sequence identifier. This modifies the identifiers in the output file to contain only the sequence identifer, discarding any comment information (which could always be obtain from the input fasta file). The formatting of the assign_taxonomy result files needs to be centralized, and at that stage this processing should happen there rather than here (see issue #543). This change is necessary for these files to work with add_metadata.py, which is the reason this is going in now.