removed basic script test functionality
this was starting to cause more trouble that it was worth as we now have much better script interface testing in place via the script usage tests and the qiime_test_data repository. see issue #557 for additional discussion.

additionally added an option to specify only certain script usage tests (similar to what the user can do for the unit tests).