re-organized the layout of the QIIME tutorials
we've got a lot of them now, so makes sense to break them apart a little bit. i'd really like to have a catch-all glob under "Other", but putting a glob in there gets all of the pages since I the layout involves multiple :toc:s. I could figure out how to add subheadings in the :toc:s,which would likely be a better solution.