started working on new method of distance matrix class get_adjacent_distances
this will be useful for plotting change in distance between adjacent timepoints, for example in a timeseries or following a spatial gradient. this isn't working right now, but need to run for a bit, so want to commit what i currently have. need to figure out indexing into distance matrices.