updated test code to suppress stderr
one of the samples in this test data set is composed of all non-16S sequences, and thus has no tips in the tree. unifrac raises a warning as a result, but because that warning is raised from within a beta_diversity.py system call, we can't suppress it with warnings.filterwarnings(). to get around this, we're essentially redirecting stderr temporarily during each test.