removed test for presence of wf.test.otu.CleanUp.ReferenceOTU0
because the subsampling is random, we occasionally don't have any "clean-up" OTUs. this test was failing too frequently for it to be useful (and running it multiple times to check for presence in a subset of the results wouldn't work well here because ofthe long run time of the test).