added prefiltering of biom table
Now removes samples with fewer than sampling_depth sequences before running any analyses. Previously samples with very few sequences (e.g., 1) were included in the rarefaction plots and taxa summaries, which was misleading (for the taxa summaries) and inconvenient (for the rarefaction plots, as some lines would end before others). This additionally ensures that the same samples are compared in all of the core diversity analyses. A link to this gzipped biom table is included in the index.html.