Removed make_cluster_jobs.py test.
Discussed with @gregcaporaso and we decided that this test should be removed since it is essentially testing that qsub runs correctly with the generated jobs file. This means that in order to get all QIIME tests to pass on a system, a queueing system has to be correctly set up, which is a pain.