Add descriptive error to add_alpha_to_mapping_file
The library function was changed to check for the existance of the
specified rarefaction depth; it will now raise an exception if a
non-existant depth is passed. This exception is now catched from the
script interface and displayed using option_parser.error.

The test suite was updated accordingly for this module.

An unnecessary type cast was also removed from the script interface.

This should fix #827