Add default depth as the highest
Mean alpha is now retrieve the highest available rarefaction depth when
None is specified. Script usage and unit tests were updated accordingly.

A new option was added to the script interface to indicate when the user
is trying to pass collated data as the input, since --depth is no longer
required to determine whether or not the execution will be over collated
or non-collated data, there was a redundancy that this option resolves.