Significant changes to script interface
Changes mades to the script interface so that the --md_fields option would be required. This, coupled with a change to the library, allows an individual to passbiom files which have jagged metadata (if metadata is list or string). Laura pointed out that this was likely the most common use case, and so it needed to be supported.