Modified compare_alpha_diversity.py
This commit modifies compare_alpha_diversity.py so that it will output the average alpha diversity of the compared categories. I find myself wanting this information frequently when doing tests and it wasn't available. These changes modify the output of the script so that you pass a directory to be created and it puts two files in that directory. One specifying the comparisons and their values, and one specifying the averages and std's of the averages for each treatment group. I have updated the script usage tests and verified that they work so that jenkins should be happy.

In addition, I needed to modify core_diversity_analysis.py because it depended on the output. I only had to modify several lines in the workflow/core_diversity_analysis.py code to alter the way it was handling the output directories.