Fixed deadlocking error.
Diagnosed a strange deadlocking error where if an invalid values in double scalers error was encountered by numpy during the kruskal wallis calculations, the interpreter would just hang rather than report an error. This SO post describes the error: http://stackoverflow.com/questions/11670404/mod-wsgi-hangs-on-runtimewarning-divide-by-zero

A simple check for 0 divisor suffices in this case, but in other cases this may be hard to detect. Some global resolution to this kind of issue would be nice.