Documentation changes and small qiime.util change
This commit is mainly about adding several hundred lines of documentation to test_group_significance.py to meet the requests that @lursell received during a recent code review.
Other changes include a slight modification to sync_biom_and_mf (and associated test files) so that it would return the non-overlapping samples as a list.