Changes to tests
This commit changes a significant number of tests in different locations. This was necessitated by finding out that there was a slight miscalculation occurring in mc_t_two_sample where the number of permutations and extreme t-stats was each being inflated by 1.

Additional changes centered on fixing the problems with the previous commit for core_diversity_analysis.py as well as adding another script usage test.