Add sanity checks to parse_coords ( fixes #982 )
Add a couple checks to the coordinates parser, doing this prevents the
parser to load distance matrices seamlessly. Specifically the function
will now check for:
    - First line being the number of vectors.
    - Second to last line being eigenvalues.
    - Last line being the percentage of variation explained.

When these conditions are not mete a `QiimeParseError` exception is
raised; the test suite was extended to check it raises these errors on
different cases.