Remove "versions" from the QIIME overview tutorial
The document was pointing to an FTP site that read "QIIME-1.5.0" and in
turn was confusing users. I've changed this such that there's now a
folder named qiime-files in thebeast.colorado.edu where I've copied the
current contents of qiime/examples/qiime_tutorial/ and updated the
documentation of tutorial.rst.

Fixes #1137