Fix #1157
Resolve the problem where OTU maps would contain duplicated observation
identifiers and in turn would cause the biom formatter to fail.

The problem was originated because the command being executed was:

    cat otu_map_1.txt otu_map_2.txt otu_map_3.txt >> full_map.txt

I've change this such that the '>>' operator is replaced with '>' hence
overwritting the file named full_map.txt in case it existed i. e. when
passing the --force flag from the script interface.