Update code for mpl 1.3.1 upgrade
Test in tests/test_stats.py was checking the x ticks of an empty plot, which apparently are different between mpl 1.1.0 and 1.3.1.

In this case, the fix is to remove the test since it is version-specific and isn't all that useful.

I manually ran through the tutorials for make_distance_boxplots.py, make_distance_comparison_plots.py, and compare_distance_matrices.py and the generated figures look good.

Any other scripts you'd like me to manually test? Switching mpl versions is always a little scary...