Changes to build denoiser
The denoiser's Makefile was specifying rm *.o for the clean rule; if
this was the case and no *.o files existed, this would error out hence
not executing the subsequent lines.

The setup.py function was copying all the files from the
denoiser/FlowgramAlignment/ folder, this now only copies the *.lhs files
and the Makefile.

Add FlowgramAli_4frame to MANIFEST.in so it's not copied on distribution
and upload time.