print_qiime_config.py mpl version check handles rc
The version check for matplotlib that is performed in print_qiime_config.py now handles version strings that end in ```rc```. For example, ```1.1.1rc``` now passes the version check instead of throwing a ```ValueError``` when casting the last element to an integer.

Ran into this issue when testing out a QIIME installation that had mpl 1.1.1rc installed.