BUG/ENH: Let write_biom_table fallback to JSON
`write_biom_table` required the availability of H5PY, this has been
changed to depend on an optional parameter that sets whetehr we want a
JSON or an HDF5 formatted BIOM file written to disk.

I've added an extra test for the case where we want to force to write as
a JSON string.

What this means is that QIIME does not require HDF5 and H5PY to be
installed, however if these two are available it will give preference to
HDF5 over JSON.

Fixes #1597
Fixes #1739