REVERT: changed default gap threshold back to 0.999999
due to a rounding error issue in scikit-bio (biocore/scikit-bio#801), the
machine-epsilon-based default value is too low, and positions that are all gaps
are not filtered out. this will be fixed in scikit-bio, but in the meantime
we're reverting to the value that had been in place in QIIME 1.8.0 and earlier.