BUG: Add --read_arguments_from_file to sl_fastq.py
With this new argument split_libraries_fastq.py can now take a file that
list the input files for the `-b`, `-i`, `--sample_ids` and
`--mapping_files` arguments. Thus allowing an indefinite amount of files
as inputs, as opposed to the previous approach where we would be limited
by the maximum number of characters that can fit into a single command
line execution.

Fixes #2069